增加购物车服务

+++ shared/Zhaoxi.MSACommerce.SharedKernel/Domain/AuditWithUserEntity.cs
@@ -2,6 +2,6 @@
 
 public abstract class AuditWithUserEntity : BaseAuditEntity
 {
-    public int? CreatedBy { get; set; }
-    public int? LastModifiedBy { get; set; }
+    public long? CreatedBy { get; set; }
+    public long? LastModifiedBy { get; set; }
 }
